claude(scan-git-for-tils): break files up into packages and smaller modules
