+++ tools/claude/config/skills/scan-git-for-tils/git/__init__.py
@@ -1,0 +1,1 @@
+"""Git and GitHub integration for TIL workflow."""
