+++ tools/claude/config/skills/scan-git-for-tils/notion/__init__.py
@@ -1,0 +1,1 @@
+"""Notion integration for TIL workflow."""
